Feat: move enum socialtype to global module (#66)
