+++ Projects/Data/Sources/RequestDTO/AuthRequestDTO.swift
@@ -15,17 +15,3 @@ public struct AuthRequestDTO: Encodable {
         self.nickName = nickName
     }
 }
-
-public enum SocialType: String {
-    case APPLE
-    case KAKAO
-    
-    public init(rawValue: String) {
-        switch rawValue {
-        case "KAKAO": self = .KAKAO
-        case "APPLE": self = .APPLE
-        default:
-            fatalError("KAKAO, APPLE이 아님")
-        }
-    }
-}
